ClipBoard OCR/requirements.txt
pytesseract==0.3.8
pynput==1.7.3
Pillow==9.0.0
